+++ BitbucketServerAutomation/Functions/Add-PSTypeName.ps1
@@ -6,7 +6,10 @@ function Add-PSTypeName
         $InputObject,
 
         [Switch]
-        $RepositoryInfo
+        $RepositoryInfo,
+
+        [Switch]
+        $ProjectInfo
     )
 
     process
@@ -14,6 +17,15 @@ function Add-PSTypeName
         if( $RepositoryInfo )
         {
             $InputObject.pstypenames.Add( 'Atlassian.Bitbucket.Server.RepositoryInfo' )
+        }
+
+        if( $ProjectInfo )
+        {
+            $InputObject.pstypenames.Add( 'Atlassian.Bitbucket.Server.ProjectInfo' )
+            if( -not ($InputObject | Get-Member -Name 'description') )
+            {
+                $InputObject | Add-Member -MemberType NoteProperty -Name 'description' -Value ''
+            }
         }
 
         $InputObject