+++ RELEASE_NOTES.md
@@ -1,1 +1,6 @@
 
+# 0.0.0
+
+ * Created `New-BBServerProject` function for creating projects.
+ * Created `Get-BBServerProject` function for getting projects.
+